Add ppt & battle scene small fix

+++ MountainTale/Assets/BattleSceneScripts/Character.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public class Character : MonoBehaviour
 {
     public string name;
-    public int MaxHP;
-    public int currHP;
+    public float MaxHP;
+    public float currHP;
     public int dmg;
     public int def;
 
     public string GetName() { return name; }
-    public int GetMaxHp() { return MaxHP; }
+    public float GetMaxHp() { return MaxHP; }
     public int GetDmg() { return dmg; }
     public int GetDef() { return def; }
 }
